# types/request.go
package types

import "mime/multipart"

type LoginRequest struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type PaginatedRequest struct {
	Page  int64 `json:"page" binding:"required"`
	Limit int64 `json:"limit" binding:"required"`
}

type MaterialsProfileFilterRequest struct {
	Sector                string   `json:"sector"`
	MaintenanceIDs        []string `json:"maintenance_ids"`
	EquipmentMachineryIDs []string `json:"equipment_machinery_ids"`
}

var (
	UPDATE_TYPE_NEW    = "new"
	UPDATE_TYPE_MODIFY = "modify"
)

type UpdateMaterialsRealityProfileRequest struct {
	UpdateType string `json:"update_type" binding:"required"`
}

type UpdateMaterialsEstimateProfileRequest struct {
}

type UpdateUserInfoRequest struct {
	FullName      string `json:"full_name"`
	Workspace     string `json:"workspace"`
	WorkspaceRole string `json:"workspace_role"`
}

type UpdateUserPasswordRequest struct {
	OldPassword string `json:"old_password" binding:"required"`
	NewPassword string `json:"new_password" binding:"required"`
}

type CreateMaterialRequestReq struct {
	MaintenanceInstanceID string                           `json:"maintenance_instance_id"`
	Sector                string                           `json:"sector" binding:"required"`
	Description           string                           `json:"description"`
	MaterialsForEquipment map[string]MaterialsForEquipment `json:"materials_for_equipment" binding:"required"`
}

type MaterialRequestUpdate struct {
	ID                    string                           `json:"id" binding:"required"`
	Sector                string                           `json:"sector" bson:"sector"`
	Description           string                           `json:"description" bson:"description"`
	MaterialsForEquipment map[string]MaterialsForEquipment `json:"materials_for_equipment" bson:"materials_for_equipment"`
}

type CreateMaintenanceRequest struct {
	Project           string `json:"project" binding:"required"`
	ProjectCode       string `json:"project_code" binding:"required"`
	MaintenanceTier   string `json:"maintenance_tier" binding:"required"`
	MaintenanceNumber string `json:"maintenance_number" binding:"required"`
}

type UploadEstimateSheetRequest struct {
	MaintenanceInstanceID string                `json:"maintenance_instance_id" binding:"required"`
	Sheet                 *multipart.FileHeader `json:"sheet"`
	SheetName             string                `json:"sheet_name" binding:"required"`
	Sector                string                `json:"sector" binding:"required"`
}

type MaterialRequestExport struct {
	MaterialRequestID string `json:"material_request_id" binding:"required"`
}

type CreateEquipmentMachineryReq struct {
	Name   string `json:"name" binding:"required"`
	Sector string `json:"sector" binding:"required"`
}

type CreateMaterialProfileReq struct {
	MaintenanceInstanceID string                `json:"maintenance_instance_id" binding:"required"`
	EquipmentMachineryID  string                `json:"equipment_machinery_id" binding:"required"`
	IndexPath             string                `json:"index_path" binding:"required"`
	Sector                string                `json:"sector" binding:"required"`
	Estimate              MaterialsForEquipment `json:"estimate" binding:"required"`
}